Move app logic to List component
