+++ onboarding-task/src/components/List.jsx
@@ -1,37 +1,66 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 
 import { ListItem } from './ListItem.jsx';
 import { ListItemInput } from './ListItemInput';
+import { generateGUID } from '../utils/GUIDGenerator';
 
-export function List(props) {
-  return (
-    <div className="row">
-      <div className="col-sm-12 col-md-6">
-        <ol className="list">
-          {
-            props.items.map(item => (
-              <ListItem
-                key={item.id}
-                id={item.id}
-                itemName={item.itemName}
-                onUpdateItem={props.onUpdateItem}
-                onDeleteItem={props.onDeleteItem}
-              />
-            ))
-          }
-        </ol>
-        <ListItemInput onCreateItem={props.onCreateItem} />
+export class List extends React.PureComponent {
+  static displayName = 'List';
+
+  constructor(props) {
+    super(props);
+    this.state = { items: [
+      { id: generateGUID(), itemName: 'Make coffee' },
+      { id: generateGUID(), itemName: 'Master React' },
+    ],
+    };
+  }
+
+  _createNewItem = (value) => {
+    this.setState(prevState => ({
+      items: prevState.items.concat({ id: generateGUID(), itemName: value }),
+    }));
+  };
+
+  _updateItem = (key, value) => {
+    const newArray = [];
+    this.state.items.forEach(item => {
+      if (item.id === key) {
+        newArray.push({ id: key, itemName: value });
+      }
+      else {
+        newArray.push(item);
+      }
+    });
+    this.setState({ items: newArray });
+  };
+
+  _deleteItem = (key) => {
+    this.setState(prevState => ({
+      items: prevState.items.filter(item => item.id !== key),
+    }));
+  };
+
+  render() {
+    return (
+      <div className="row">
+        <div className="col-sm-12 col-md-6">
+          <ol className="list">
+            {
+              this.state.items.map(item => (
+                <ListItem
+                  key={item.id}
+                  id={item.id}
+                  itemName={item.itemName}
+                  onUpdateItem={this._updateItem}
+                  onDeleteItem={this._deleteItem}
+                />
+              ))
+            }
+          </ol>
+          <ListItemInput onCreateItem={this._createNewItem} />
+        </div>
       </div>
-    </div>
-  );
+    );
+  }
 }
-
-List.displayName = 'List';
-
-List.propTypes = {
-  items: PropTypes.array.isRequired,
-  onCreateItem: PropTypes.func.isRequired,
-  onUpdateItem: PropTypes.func.isRequired,
-  onDeleteItem: PropTypes.func.isRequired,
-};
